2.0 API updates and bug fixes (#1180)
* Implemented DatabaseConfiguration(Context context) for Android (part of #1174)

* Implemented Dictionary.getKeys() instead of getAllKeys().  (Part of #1174)

* Removed Database(String name) constructor for Java/Andorid platforms

* Implemented Query.explain() (This is part of #1174)

* Added two unit tests for DocumentChange Listener

* Implemented testRemoveDictionary() unit test which was not ported. And also fixed the issue which was found by the test.

* Fixed #1158 - 2.0: getLong/setLong does not work as expected
  - Casting long value to int causes data loss. Main fix is in litecore
  - Update unit tests
  - Replaced HashMap with TreeMap in Unit Test because TreeMap can guarantee key order.
  - Make Dictionary.toMap() uses TreeMap as implementation instead of HashMap.

* Fixed #1160 - 2.0: Double/Float without decimal are returned as Integer
  - Fleece stores Double/Float value without decimal fraction as Integer. Return value type will be Long from the Database.

* Fixed #1161 - 2.0: Double value which is stored in fleece is returned as Float.
  - Fleece stores Double value as Float if double value is same with float value.

* Fixed #1170 - Array and Dictionary should throw ConcurrentModificationException if Array or Dictionary is modified.
  - Array: As all data are loaded into the internal List, so just rely on Java implementation.
  - Dictionary: Like java.util.AbstractMap implementation, used expectChangeCount technique to detect modification during iteration.

* Switched back from TreeMap to HashMap
  - Fixed Map value comparison in unit test.
   - All integral values from Fleece are Long. Integer and Long values companions fails.

* Fixed DatabaseConfiguration constructor and Dictionary.getKeys()
  - DatabaseConfiguration have one constructor with Android context
  - (ReadOnly)Dictionary.getKeys() creates copy of List<String>.

* Fixed double creation of ArrayList instance

+++ shared/src/main/java/com/couchbase/lite/SharedKeys.java
@@ -71,7 +71,7 @@ public class SharedKeys implements FLValue.ISharedKeys {
             return null;
 
         if (key.isInteger())
-            return getKey(key.asInt());
+            return getKey((int)key.asInt());
 
         return (String) valueToObject(key);
     }
